+++ Cargo.toml
@@ -12,6 +12,7 @@ members = [
   "quickwit-doc-mapper",
   "quickwit-indexing",
   "quickwit-ingest-api",
+  "quickwit-janitor",
   "quickwit-metastore",
   "quickwit-proto",
   "quickwit-search",
