+++ quickwit-janitor/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "quickwit-janitor"
+version = "0.3.1"
+authors = ["Quickwit, Inc. <hello@quickwit.io>"]
+edition = "2021"
+license = "AGPL-3.0-or-later" # For a commercial, license, contact hello@quickwit.io
+description = "Quickwit indexing"
+repository = "https://github.com/quickwit-oss/quickwit"
+homepage = "https://quickwit.io/"
+documentation = "https://quickwit.io/docs/"
+
+[dependencies]
+anyhow = "1"
+async-trait = "0.1"
+quickwit-actors = { version = "0.3.1", path = "../quickwit-actors" }
+quickwit-config = { version = "0.3.1", path = "../quickwit-config", features = [
+  "testsuite"
+] }
+quickwit-metastore = { version = "0.3.1", path = "../quickwit-metastore" }
+quickwit-storage = { version = "0.3.1", path = "../quickwit-storage" }
+serde = "1"
+tokio = { version = "1", features = ["sync"] }
+tracing = "0.1.29"
+
+[dev-dependencies]
+quickwit-common = { version = "0.3.1", path = "../quickwit-common" }
+quickwit-metastore = { version = "0.3.1", path = "../quickwit-metastore", features = [
+    "testsuite",
+] }
+quickwit-storage = { version = "0.3.1", path = "../quickwit-storage", features = [
+    "testsuite",
+] }
+tempfile = "3"
