+++ quickwit-janitor/src/lib.rs
@@ -1,0 +1,46 @@
+// Copyright (C) 2022 Quickwit, Inc.
+//
+// Quickwit is offered under the AGPL v3.0 and as commercial software.
+// For commercial licensing, contact us at hello@quickwit.io.
+//
+// AGPL:
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+use std::sync::Arc;
+
+use actors::JanitorService;
+use quickwit_actors::{Mailbox, Universe};
+use quickwit_config::QuickwitConfig;
+use quickwit_metastore::Metastore;
+use quickwit_storage::StorageUriResolver;
+use tracing::info;
+
+pub mod actors;
+
+pub async fn start_janitor_service(
+    universe: &Universe,
+    config: &QuickwitConfig,
+    metastore: Arc<dyn Metastore>,
+    storage_uri_resolver: StorageUriResolver,
+) -> anyhow::Result<Mailbox<JanitorService>> {
+    info!("Starting janitor service.");
+    let janitor_service = JanitorService::new(
+        config.data_dir_path.to_path_buf(),
+        metastore,
+        storage_uri_resolver,
+    );
+    let (janitor_service_mailbox, _) = universe.spawn_actor(janitor_service).spawn();
+
+    Ok(janitor_service_mailbox)
+}
