+++ quickwit-serve/Cargo.toml
@@ -35,6 +35,7 @@ quickwit-directories = { version = "0.3.1", path = "../quickwit-directories" }
 quickwit-doc-mapper = { version = "0.3.1", path = "../quickwit-doc-mapper" }
 quickwit-indexing = { version = "0.3.1", path = "../quickwit-indexing" }
 quickwit-ingest-api = { version = "0.3.1", path = "../quickwit-ingest-api" }
+quickwit-janitor = { version = "0.3.1", path = "../quickwit-janitor" }
 quickwit-metastore = { version = "0.3.1", path = "../quickwit-metastore" }
 quickwit-proto = { version = "0.3.1", path = "../quickwit-proto" }
 quickwit-search = { version = "0.3.1", path = "../quickwit-search" }
