+++ quickwit-serve/src/lib.rs
@@ -48,9 +48,11 @@ use quickwit_cluster::{Cluster, ClusterMember, QuickwitService};
 use quickwit_common::uri::Uri;
 use quickwit_config::QuickwitConfig;
 use quickwit_core::IndexService;
-use quickwit_indexing::actors::{IndexingService, JanitorService};
-use quickwit_indexing::{start_indexing_service, start_janitor_service};
+use quickwit_indexing::actors::IndexingService;
+use quickwit_indexing::start_indexing_service;
 use quickwit_ingest_api::{start_ingest_api_service, IngestApiService};
+use quickwit_janitor::actors::JanitorService;
+use quickwit_janitor::start_janitor_service;
 use quickwit_metastore::{quickwit_metastore_uri_resolver, Metastore, MetastoreGrpcClient};
 use quickwit_search::{start_searcher_service, SearchService};
 use quickwit_storage::quickwit_storage_uri_resolver;
